+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - promisify index.js logic
+- move logic from index.js to main.js in `server` folder
 
 ## [v0.2.0] - 2020-05-23
 ### Added